requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
requests==2.31.0
python-dotenv==1.0.0
pydantic==1.10.12
pydantic-core==0.18.1
aiohttp==3.9.1
beautifulsoup4==4.12.2
markdown==3.5.1
